pyproject.toml
[project]
name = "a4a-lab"
version = "0.2.20"
description = "This project is A4A(Agent for Agent) of T2A(Text to Agent) and aim to CoA(Chain of AI) on AOTL(AI On The Loop)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[a2a]>=1.21.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.132.0",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "python-dotenv>=1.2.1",
    "line-bot-sdk>=3.0.0",
    "uvicorn>=0.27.0",
    "fastapi>=0.109.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
a4a = "a4a_lab.run:main"

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["src"]
